+++ TablePro/Views/Sidebar/Menu/SidebarMenuCommand.swift
@@ -35,6 +35,8 @@ internal enum SidebarMenuCommand: Equatable {
     case removeRecent(DatabaseTreeTableRef)
     case clearRecents
     case useAsActive(DatabaseContainerRef)
+    case setFavoriteDatabase(database: String, environment: FavoriteDatabaseEnvironment)
+    case removeFavoriteDatabase(String)
     case refreshContainers([DatabaseContainerRef])
     case copyContainerNames([DatabaseContainerRef])
     case exportContainers([DatabaseContainerRef])